Refactor App.js and playlist data
This refactor optimizes the code in App.js by improving the
 autoplay logic. Additionally, the authorId
field in the playlist data has been updated to be an array instead of a
string. This change allows for multiple authorIds to be associated with
a playlist. The modifications enhance code efficiency and provide
flexibility for playlist authorship. Fixes #1234